Removes compiler warnings (unused variables) and signed/unsigned comparison

+++ include/bitcoin/bitcoin/impl/machine/program.ipp
@@ -101,7 +101,7 @@ inline bool program::increment_operation_count(const operation& op)
 inline bool program::increment_multisig_public_key_count(int32_t count)
 {
     // bit.ly/2d1bsdB
-    if (count < 0 || count > max_script_public_keys)
+    if (count < 0 || count > static_cast<int32_t>(max_script_public_keys))
         return false;
 
     // Addition is safe due to script size validation.
